requirements.txt
python-telegram-bot
flask
supabase
flask[async]
requests
